requirements.txt
bcrypt>=4.0.1
fastapi>=0.99.1
uvicorn>=0.22.0
langchain>=0.0.173
pinecone-client==2.2.1
python-dotenv>=1.0.0
python-multipart>=0.0.6
openai>=0.27.6
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
pydantic[email]>=1.10.7
requests
chromadb
google-auth>=2.21.0
oauth2client>=4.1.3
authlib>=1.2.1
itsdangerous>=2.1.2
httpx>=0.17.3
pytest
httpx
jsonschema
tiktoken
